requirements.txt
google-generativeai>=0.8.0
streamlit>=1.25.0
fpdf>=1.7.2
pdf2image>=1.16.0
Pillow>=9.0.0
# Note: tkinter comes with Python standard library, no separate installation needed 